Creates capybara_ext.rb and corrects capybara.rb and spec_helper.rb
